requirements.txt
oracledb>=3.0.0
reportlab>=4.0.0
Flask
python-dotenv